+++ tests/testthat/test-generators.R
@@ -20,26 +20,17 @@ test_that("PCG64: setting seed produces identical uniformly distributed numbers"
   expect_equal(u1, u2)
 })
 
-test_that("Xorshift128+: setting seed produces identical uniformly distributed numbers", {
-  dqRNGkind("Xorshift128+")
-  dqset.seed(seed)
-  u1 <- dqrunif(10)
-  dqset.seed(seed)
-  u2 <- dqrunif(10)
-  expect_equal(u1, u2)
-})
-
-test_that("Xorshift1024*: setting seed produces identical uniformly distributed numbers", {
-  dqRNGkind("Xorshift1024*")
-  dqset.seed(seed)
-  u1 <- dqrunif(10)
-  dqset.seed(seed)
-  u2 <- dqrunif(10)
-  expect_equal(u1, u2)
-})
-
 test_that("Xoroshiro128: setting seed produces identical uniformly distributed numbers", {
   dqRNGkind("Xoroshiro128+")
+  dqset.seed(seed)
+  u1 <- dqrunif(10)
+  dqset.seed(seed)
+  u2 <- dqrunif(10)
+  expect_equal(u1, u2)
+})
+
+test_that("Xoshiro256: setting seed produces identical uniformly distributed numbers", {
+  dqRNGkind("Xoshiro256+")
   dqset.seed(seed)
   u1 <- dqrunif(10)
   dqset.seed(seed)